Исправил структуру проекта Исправил докер (все еще требует предварительно сборки) Добавил докер файл для crm Исправил ошибку кролика (не создавал очередь crm.queue)
