--- crm/docker-compose.yaml
@@ -1,21 +1,0 @@
-version: "3.8"
-
-services:
-  postgres:
-    image: postgres:16-alpine
-    container_name: crm_postgres_db
-    volumes:
-      - postgres-data:/var/lib/postgresql/data
-    environment:
-      POSTGRES_USER: postgres
-      POSTGRES_PASSWORD: postgres
-      POSTGRES_DB: crmdb
-    ports:
-      - "5434:5432"
-
-networks:
-  microservices-network:
-    driver: bridge
-
-volumes:
-  postgres-data:
