--- src/main/java/com/projects/nexi/NexiApplication.java
@@ -1,13 +1,0 @@
-package com.projects.nexi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NexiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(NexiApplication.class, args);
-    }
-
-}
