--- src/test/java/com/projects/nexi/NexiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.projects.nexi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NexiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
